requirements.txt
requests
moviepy
edge-tts
python-dotenv
